terminando de desenvolver a API

+++ src/main/java/com/projectmanager/service/ProjectManagerService.java
@@ -5,12 +5,13 @@ import com.projectmanager.database.postgresql.dao.ProjetoDAO;
 import com.projectmanager.model.Pessoa;
 import com.projectmanager.model.Projeto;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 import org.springframework.stereotype.Service;
 
 @Service("ProjectManager")
 public class ProjectManagerService {
+    
+    // Table: projectmanager.pessoa
     
     public long create(Pessoa pessoa) throws SQLException {
         Pessoa p = PessoaDAO.get(pessoa);
@@ -30,15 +31,17 @@ public class ProjectManagerService {
         return PessoaDAO.get(Long.parseLong(id));
     }
     
-    public List<Pessoa> getPessoas(long page, long limit) throws SQLException {
+    public List<Pessoa> getPessoas(String coluna, String ordem, long page, long limit) throws SQLException {
         long offset = (page - 1) * limit;
-        return PessoaDAO.list(limit, offset);
+        return PessoaDAO.list(coluna, ordem, limit, offset);
     }
     
-    public List<Pessoa> buscarPessoas(String q, long page, long limit) throws SQLException {
+    public List<Pessoa> buscarPessoas(String q, String coluna, String ordem, long page, long limit) throws SQLException {
         long offset = (page - 1) * limit;
-        return PessoaDAO.find(q, limit, offset);
+        return PessoaDAO.find(q, coluna, ordem, limit, offset);
     }
+    
+    // Table: projectmanager.projeto
     
     public long create(Projeto projeto) throws SQLException {
         return ProjetoDAO.insert(projeto);
@@ -49,36 +52,51 @@ public class ProjectManagerService {
     }
     
     public int delete(Projeto projeto) throws SQLException {
-        return ProjetoDAO.delete(projeto);
+        if (!projeto.getStatus().matches("iniciado|em andamento|encerrado")) return ProjetoDAO.delete(projeto);
+        else return 0;
     }
     
     public Projeto getProjeto(String id) throws SQLException {
         return ProjetoDAO.get(Long.getLong(id));
     }
     
-    public int associate(Projeto projeto, Pessoa pessoa) {
-        return 0;
-    }
-    
-    public int disassociate(Projeto projeto, Pessoa pessoa) {
-        return 0;
-    }
-    
-    public List<Pessoa> getMembers(long page, long limit) {
-        return new ArrayList();
-    }
-    
-    public List<Pessoa> findMembers(String q, long page, long limit) {
-        return new ArrayList();
-    }
-    
-    public List<Projeto> getProjetos(long page, long limit) throws SQLException {
+    public List<Projeto> getProjetos(String coluna, String ordem, long page, long limit) throws SQLException {
         long offset = (page - 1) * limit;
-        return ProjetoDAO.list(limit, offset);
+        return ProjetoDAO.list(coluna, ordem, limit, offset);
     }
     
-    public List<Projeto> buscarProjetos(String q, long page, long limit) throws SQLException {
+    public List<Projeto> buscarProjetos(String q, String coluna, String ordem, long page, long limit) throws SQLException {
         long offset = (page - 1) * limit;
-        return ProjetoDAO.find(q, limit, offset);
+        return ProjetoDAO.find(q, coluna, ordem, limit, offset);
+    }
+    
+    // Table: projectmanager.membros
+    
+    public long associate(long idProjeto, long idPessoa) throws SQLException {
+        return ProjetoDAO.adicionarMembro(idProjeto, idPessoa);
+    }
+    
+    public int disassociate(long idProjeto, long idPessoa) throws SQLException {
+        return ProjetoDAO.removerMembro(idProjeto, idPessoa);
+    }
+    
+    public List<Pessoa> getMembers(long projetoId, String coluna, String ordem, long page, long limit) throws SQLException {
+        long offset = (page - 1) * limit;
+        return ProjetoDAO.listarMembros(ProjetoDAO.get(projetoId), coluna, ordem, limit, offset);
+    }
+    
+    public List<Projeto> getProjects(long pessoaId, String coluna, String ordem, long page, long limit) throws SQLException {
+        long offset = (page - 1) * limit;
+        return ProjetoDAO.listarProjetos(PessoaDAO.get(pessoaId), coluna, ordem, limit, offset);
+    }
+    
+    public List<Pessoa> findMembers(String q, long projetoId, String coluna, String ordem, long page, long limit) throws SQLException {
+        long offset = (page - 1) * limit;
+        return ProjetoDAO.buscarMembros(q, ProjetoDAO.get(projetoId), coluna, ordem, limit, offset);
+    }
+    
+    public List<Projeto> findProjects(String q, long pessoaId, String coluna, String ordem, long page, long limit) throws SQLException {
+        long offset = (page - 1) * limit;
+        return ProjetoDAO.buscarProjetos(q, PessoaDAO.get(pessoaId), coluna, ordem, limit, offset);
     }
 }
